ci: clone all repos during kubesaw deployment
During the deployment process for kubesaw, toolchain-e2e needs to have
access to host-operator and member-operator repositories to render
deployment templates.  If the build system isn't given an explicit path
to a tree, it'll clone off of whatever's sitting in latest trunk.  Since
we're running a fork of it, we need to tell toolchain-e2e to use our
forks as well.

Signed-off-by: Andy Sadler <ansadler@redhat.com>