ci: run more tests more often
If we want to make the unit-testing job always pass before we can merge
a PR, we need to have it run in almost all circumstances. With this
change, all PRs will have unit tests run, not just when relevant paths
have changed.

This also sets up unit testing for operator/ and fixes a bug where we
were running the latest version of golang-ci instead of the pinned
version.

Signed-off-by: Andy Sadler <ansadler@redhat.com>