add release components
This adds the necessary infrastructure to make a draft release.
Currently, artifacts will be the config/ directories of both components
wrapped up in tarballs, with their images pointing to the release build.
We can add more to our artifacts if we need to in the future.

Creating a release will be done using a workflow dispatch, which takes
the tag to be created as an input.

Signed-off-by: Andy Sadler <ansadler@redhat.com>