add release components (#199)
* add release components

This adds the necessary infrastructure to make a draft release.
Currently, artifacts will be the config/ directories of both components
wrapped up in tarballs, with their images pointing to the release build.
We can add more to our artifacts if we need to in the future.

Creating a release will be done using a workflow dispatch, which takes
the tag to be created as an input.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* write manifest tarballs to out/ directories

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* ignore contents of out/ directories

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* parameterize JWKS_URL

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* remove trailing * from gitignores

Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>
Co-authored-by: Francesco Ilario <filario@redhat.com>