operator: update config for a newer kustomize
By default, when operator-sdk scaffolds a new operator, it uses a fairly
old dialect of kustomize's configuration language, which isn't supported
in newer versions of kustomize.  If we want to be consumed into
infra-deployments directly, we need to update our configs to have
compatibility with kustomize v5.0 and later.

While we're at it, we're not likely going to be using OLM as a
deployment mechanism, so keeping support for it is just bloat in our
repository.  It also results in more things landing in our release that
we don't really need (such as samples and a scorecard).  Let's remove
this stuff to prune things.

Signed-off-by: Andy Sadler <ansadler@redhat.com>