add kube-linter checks
Infra-deployments has a suite of checks that makes sure every kustomize
manifest committed passes kube-linter in its default configuration.  If
we want to release onto platforms supported by infra-deployments, we
need to be able to pass these checks too.

This introduces a new check in CI to ensure that we don't fall
out-of-line with what infra-deployments expects.

Signed-off-by: Andy Sadler <ansadler@redhat.com>