introduce The Workspaces Book
This commit adds a first draft version of The Workspaces Book.
MDBook is used for building the book.

This commit also adds a GitHub Action responsible to build and publish
the book to GitHub Pages.

Support to Vale is also added. Vale can be used to lint the documentation
and spot typos. No automation is added to automatically lint documentation,
as it may be a little overwhelming in this phase of the project.

The developer can leverage on make target for building and linting the documentation.
To simplify the developer experience and avoid them to install the tools, containerized
versions of the tools are used.

Signed-off-by: Francesco Ilario <filario@redhat.com>