add kube-linter checks (#214)
* add kube-linter checks

Infra-deployments has a suite of checks that makes sure every kustomize
manifest committed passes kube-linter in its default configuration.  If
we want to release onto platforms supported by infra-deployments, we
need to be able to pass these checks too.

This introduces a new check in CI to ensure that we don't fall
out-of-line with what infra-deployments expects.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* operator: server: set images during packaging

Rather than pre-commit our image tags (which will likely get out-of-date
very quickly), add in references to the built images during the
packaging process.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* fail kubelinter checks if warnings were emitted

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* move kubelinter checks to .github/workflows/security.yaml

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* ci: don't pretty-print sarif report on failure

this information is already captured in github's security report, so it
doesn't add anything useful.

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* fix e2e deployments

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* server: don't print all commands during deployment

this was a debugging change that shouldn't have been committed

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* address feedback

Signed-off-by: Andy Sadler <ansadler@redhat.com>

* remove trailing space

don't know how that snuck in

Signed-off-by: Andy Sadler <ansadler@redhat.com>

---------

Signed-off-by: Andy Sadler <ansadler@redhat.com>