tests: add script for smoke testing (#276)
To run these tests, you'll need to obtain a token corresponding to the
environment you're wanting to test.  Tokens are not portable, i.e. a
token from staging-external won't work for production-external.

Internal environments are not yet supported, since the PR for enabling
them has not yet landed.  See [#4225] for details.

[#4225]: https://github.com/redhat-appstudio/infra-deployments/pull/4225

Signed-off-by: Andy Sadler <ansadler@redhat.com>