add workspaces/operator in dependabot for indirect deps (#284)
This change wants to force the update of `e2e`'s indirect dependencies
related to ginkgo and gomega when they are updated in `operator`.

Look at #282 and #283 for examples of the problem. When gomega or ginkgo
are updated in `operator`, they are not updated in `e2e` as they are
indirect dependencies. As a consequence the build fails due to untidied
gomod.

Alternatively, we can add ginkgo and gomega as `e2e`'s direct dependencies.
I think it would be good to start using gomega `e2e`, but I don't see
space for ginkgo at the moment.

Signed-off-by: Francesco Ilario <filario@redhat.com>