remove kube-rbac-proxy as a dependency (#288)
We need to set up a ServiceMonitor for our service.  We could use
kube-rbac-proxy, but that has apparently been deprecated by kubebuilder,
and they recommend moving away from it.  Instead, we can do the
authentication and authorization checks inside our metrics server.

To do this, we need to set up the ServiceMonitor with a ServiceAccount
with permissions to talk to `/metrics` from our metrics service.

Signed-off-by: Andy Sadler <ansadler@redhat.com>