add is-owner label
When requesting workspaces through our server, we will now label them
with the label `internal.workspaces.konflux-ci.dev/is-owner`.  This
label will either be set to `true` or `false` depending on whether the
requesting user is an owner or not.

This label is set for all operations - get, list, create, and update.

Signed-off-by: Andy Sadler <ansadler@redhat.com>