operator: don't generate data field for metrics-reader secret
ArgoCD gets thrown into an out-of-sync loop when the data field exists
on the metrics-reader secret.  This is because the service account
controller fills in the data field with all the information a service
account secret needs, which conflicts with the empty data that ArgoCD
expects.  This data field exists because kustomize's secret generator
provides a data field by default, and it appears there isn't a way to
disable this.

To resolve this, we need to move away from using the secret generator in
our kustomization manifests and instead spell out the secret in our
manifests directly.

Signed-off-by: Andy Sadler <ansadler@redhat.com>